WebSocket Dummy arguments, prompt plans
