+++ src/widgets/mod.rs
@@ -4,4 +4,4 @@ pub mod heart_rate;
 pub mod heart_rate_display;
 pub mod info_table;
 pub mod inspect_overlay;
-pub mod save_prompt;
+pub mod prompts;
